*** empty log message ***

+++ models/package/calculatevariance.hh
@@ -52,17 +52,17 @@ class CalculateVariance : public Element {
 
     class CounterEntry {
 	public:
-	CounterEntry():pkt_sum_interval(0),pkt_sum(0),pkt_sum_sq(0),pkt_count(0),byte_count(0) {}
+	CounterEntry():pkt_sum_interval(0),pkt_sum(0),pkt_sum_sq(0.0),pkt_count(0),byte_count(0) {}
 	double get_pkt_variance(unsigned num_int) const {
 	    assert(num_int>0);
 	    double tmp_mean_sqr = (double) pkt_sum/num_int;
 	    tmp_mean_sqr = tmp_mean_sqr * tmp_mean_sqr;
-	    return ((double)pkt_sum_sq/num_int) - tmp_mean_sqr;
+	    return (pkt_sum_sq/num_int) - tmp_mean_sqr;
 	}
 
 	unsigned pkt_sum_interval; //number of packets accumulated in the current (unfinished) interval
 	unsigned pkt_sum; // Sum(X) where X is ther number of packets in the previous intervals.
-	unsigned pkt_sum_sq; //Sum(X^2) where X is ther number of packets in the previous intervals. var(X) = E(X^2) - E(X)^2;
+	double pkt_sum_sq; //Sum(X^2) where X is ther number of packets in the previous intervals. var(X) = E(X^2) - E(X)^2;
 	uint32_t pkt_count;
 	unsigned byte_count;
